Add snapshot compression
