+++ commit/commit.go
@@ -3,6 +3,11 @@
 
 package commit
 
+// Writer represents a contract that a commit writer must implement
+type Writer interface {
+	Write(commit Commit) error
+}
+
 // --------------------------- Commit ----------------------------
 
 // Commit represents an individual transaction commit. If multiple chunks are committed
@@ -21,4 +26,18 @@ func (c *Commit) Clone() (clone Commit) {
 		}
 	}
 	return
+}
+
+// --------------------------- Channel ----------------------------
+
+var _ Writer = new(Channel)
+
+// Channel represents an impementation of a commit writer that simply sends each commit
+// into the channel.
+type Channel chan Commit
+
+// Write clones the commit and writes it into the writer
+func (w *Channel) Write(commit Commit) error {
+	*w <- commit.Clone()
+	return nil
 }
