+++ commit/commit_test.go
@@ -4,6 +4,8 @@
 package commit
 
 import (
+	"io"
+	"sync/atomic"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -21,4 +23,26 @@ func TestCommitClone(t *testing.T) {
 
 	clone := commit.Clone()
 	assert.EqualValues(t, commit, clone)
+}
+
+func TestWriterChannel(t *testing.T) {
+	w := make(Channel, 1)
+	w.Write(Commit{
+		Chunk: 123,
+	})
+
+	out := <-w
+	assert.Equal(t, 123, int(out.Chunk))
+}
+
+type limitWriter struct {
+	value uint32
+	Limit int
+}
+
+func (w *limitWriter) Write(p []byte) (int, error) {
+	if n := atomic.AddUint32(&w.value, uint32(len(p))); int(n) > w.Limit {
+		return 0, io.ErrShortBuffer
+	}
+	return len(p), nil
 }
